requirements.txt
chromadb>=0.4.0
openai>=1.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
pytest>=7.4.0
pandas>=2.0.0
